Refactored graph related code
